Note: not intended for release
This commit includes the rough draft of the UI and XAML theme for the program. Saving and opening saved recipes is functional. Notes on a recipe are part of the recipe’s save-file. The calculator is not implemented yet.

Update Path/Plans: the next commit should include a functional implementation of the calculator feature, and the ability to set the density of your liquids (similar to the original program’s Grams Set feature).
That commit will be followed up with rigorous bug testing, after which the first official release will be deployed.

The ultimate aim for this solution will be an IoT analytic scale that will eventually mix a user’s recipe batches using automated pipettes and accurate load cells. The product will be available as a pre-built commodity for public purchase, however this repository comes with an MIT license so feel free to DIY.