This commit is only being made because some (working) refactoring was performed. Other changes result in known bugs... Made Grams public. ATTEMPTED to implement the calculation of grams by using a variable from GramsPage (I'm aware that this attempt won't work because it's getting a new instance of GramsPage which doesn't contain the values needed).
However, I realized a different bug that WILL occur from this approach: CalculatePage's _grams member will always be null because the GramsPage is not prepared here.
Suffice to say, I'd already thought of an alternative approach to this feature before I tried this current one, it will be easier to load the json file and deserialize it in the CalculatePage.